Fix find_files for particular filesystems on linux.
The current implementation uses readdir() and struct dirent->d_type
field for determining whether we're dealing with a file or a directory.
The manual states that the d_type field is not supported by all
filesystem types. Therefore, whenever a d_type == DT_UNKNOWN is reported,
fallback to using the fstatat() system call.